some organization in ai chat cog
